+++ example/lib/xwidget/icon_spec.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+// ignore_for_file: unused_element
+
+const Icons? _icons = null;